Convert init process to data file usage
Typing the descriptions on the command line got old, fast. Now running the
script with a single argument, the repo name checks for the data file, and
builds a blank template for it if not found. If the data file exists, it is
used to build the needed files for the new project repository.

The data file, with prompt comments built in, holds all the data from the old
command line usage, and lots more. Including a separation of the description,
the HTML <title>, into component parts, places to add the HTML and CSS supplied
by the course in the forked CodePen pens, and a place to add the pen ID for my
working copy of the project.

After creating the set of files, the data file is also expanded by the script to
supply the snippets needed to update the README files in the main repo to add
the current project to the running journal.

Also removed the CHANGELOG files from the repo. They were really nothing but
extra noise. The "changes" in this repo are nothing more than updates of the
status of projects, including adding new ones. The README files track the
addition of projects, and the git log of each project tracks their status, in
better timing than this repo could.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>