Update status of homework 4 to final
In addition, as it looks like the instructional team is not reviewing, going to
comment on the non-homework projects, I'm going to consider them as 'Final' when
the class period for the following lesson is completed. At the end of lesson 8
any projects from lesson 7 will be considered as 'Final' and those of lesson 8
are so marked after lesson 9's session.

Soon we will be working on a final project, and it will likely be a single repo
with a continually updated status. Actual usage remains to be determined.

Removed Lesson 4 Exerise 1: Nav With Flexbox as it was incorporated into a
homework assignment later anyway.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>